remove old

+++ dataquality.py
@@ -1,9 +1,11 @@
-# -*- encoding: utf-8 -*-
 
 import pandas as pd
 import numpy as np
+import scipy.stats as st
 import os
 from sklearn import preprocessing
+from sklearn.preprocessing import *
+from sklearn.feature_selection import *
 
 
 class DataQuality :
@@ -34,7 +36,7 @@ class DataQuality :
         if output != "" :
             st1.to_csv(output)
         else:
-            print st1
+            print(st1)
         return st1
 
     #离散值分析每个值的分布
@@ -44,14 +46,20 @@ class DataQuality :
             df1["field_name"] = X
             df1.iloc[0:100].to_csv(output)
         elif isinstance(X,list):
-            os.remove(output)
+            if os.path.exists(output):
+                os.remove(output)
             for col_name in X:
-                df1 = self.df_data[col_name].value_counts(ascending=False)[0:100]
+                total_count=len(self.df_data.index)
+                df1 = pd.DataFrame({"count":
+                    self.df_data[col_name].value_counts(ascending=False)[0:100]})
+                df1['frequency']=df1['count']/total_count
+                df1['total_freq']=df1['frequency'].cumsum()
+                #print df1.index
                 df1["field_name"]=col_name
                 df1.to_csv(output,mode="a")
-                #print type(df1)
+               # print type(df1)
         else:
-            print "Error X entered"
+            print ("Error X entered")
         pass
 
 
@@ -83,6 +91,8 @@ class DataQuality :
         r[r > edge_up] = edge_up
         r[r < edge_low] = edge_low
         return r
+#TODO start from here
+#migrate from excel code
 
     #采用第一种方法进行logistic变换
     def T_logistic1(self,X):
@@ -91,25 +101,69 @@ class DataQuality :
     def T_logistic2(self,X):
         pass
 
-    def T_centralize(self,X):
+    # 中心化,将原始数组换成相对中心的比例
+    def T_centralize(self, X):
+        if isinstance(X, str):
+            df1 = pd.DataFrame(self.df_data[X] - self.df_data[X].mean())
+            df1.columns = df1.columns + '_centralized'
+            return df1
+        elif isinstance(X, list):
+            df1 = self.df_data[X] - self.df_data[X].mean()
+            df1.columns = df1.columns + '_centralized'
+            return df1
+        else:
+            print("Error X entered")
         pass
 
-    def T_percentile(self,X):
+     # 分位线，将袁术数据集转换成百分比
+    def T_percentile(self, X='*'):
+        if X == '*':
+            df5 = self.df_data.rank() / self.df_data.count()
+            df5.columns = df5.columns + '_precent'
+            return df5
+        elif isinstance(X, str):
+            # df1 = pd.DataFrame(self.df_data[X]-self.df_data[X].mean())
+            df5 = pd.DataFrame(self.df_data[X].rank() / self.df_data[X].count())
+            df5.columns = df5.columns + '_precent'
+            return df5
+        elif isinstance(X, list):
+            df5 = self.df_data[X].rank() / self.df_data[X].count()
+            df5.columns = df5.columns + '_precent'
+            return df5
+        else:
+            print("Error X entered")
+
         pass
 
-    #根据频数转换连续值到离散值
-    def bin_by_frequency(self,X):
+        # 根据频数转换连续值到离散值
+    def bin_by_frequency(self, X, k=4):
+        # k = 4
+        if isinstance(X, str):
+            d1 = pd.cut(self.df_data[X], k, labels=list(range(k)))  # 等宽离散化，各个类比依次命名为0,1,2,3
+            return d1
+        else:
+            print("Error X entered")
         pass
 
-    # 根据距离转换连续值到离散值
-    def bin_by_distance(self,X):
+        # 根据距离转换连续值到离散值
+    def bin_by_distance(self, X, k=4):
+        # 等频率离散化
+        if isinstance(X, str):
+            w = [1.0 * i / k for i in range(k + 1)]
+            w = self.df_data[X].describe(percentiles=w)[4:4 + k + 1]  # 使用describe函数自动计算分位数
+            w[0] = w[0] * (1 - 1e-10)
+            d2 = pd.cut(self.df_data[X], w, labels=list(range(k)))
+            return d2
+        else:
+            print("Error X entered")
         pass
 
     #转换离散代码成数值
     def T_WOE(self,X, Y):
         pass
 
-    def AR(self,X,Y):
+    #转换离散代码成AR数值
+    def T_AR(self,X,Y):
         pass
 
     def IV(self,X,Y):
@@ -121,10 +175,40 @@ class DataQuality :
     def compute_ar(self,X,Y):
         pass
 
+#TODO 增加亚变量
+
+
     def draw_ar(self,X,Y):
         pass
 
+    def F_chi_square(self,X,Y):
+        #此处Y必须为0-好客户，1-坏客户
+        #X可以为单个字符串，或者字符串的list
+        chi_list=[]
+        if isinstance(X,str):
+            x_columns = [X]
+        else :
+            x_columns = list(X)
+        for x1 in x_columns:
+            #生成交叉表，结果f1[0]为好客户，f1[1]为坏客户
+            f1 = pd.crosstab(self.df_data[x1],self.df_data[Y])
+            #total列为总客户数，expected_bad为期望的坏客户数
+            f1['total']=f1[0]+f1[1]
+            total_prob = f1[1].sum()*1.0/f1['total'].sum()
+            f1['expected_bad']=total_prob*f1['total']
+            #调用函数计算卡方系数
+            chi=st.chisquare(f1[1],f1['expected_bad'])
+            chi_list.append([x1,chi.statistic,chi.pvalue])
+        return chi_list
+
+
 dq = DataQuality()
 dq.load_file("LoanStats_clean.csv")
-dq.dq_Analyze("test1.csv")
+#x=dq.df_data["loan_amnt"]
+#y=dq.df_data['loan_status1']
+#SelectKBest(chi2,k=1).fit_transform(x,y)
+#dq.dq_Analyze("test1.csv")
 #dq.dq_Frequency(["grade","loan_amnt"],"test2.csv")
+
+#TODO s1=np.random.rand(10)
+print(dq.F_chi_square(["grade","term"],"loan_status1"))
